requirements.txt
feedparser>=6.0
httpx>=0.27
apscheduler>=3.10,<4
pyyaml>=6.0
beautifulsoup4>=4.12
lxml>=5.0
